# ENH: Improve performance of np.tensordot