In this PR we improve the performance of `np.tensordot` for small arrays by using the python builtin `math.prod` (available since python 3.8).

Benchmark:
```
import numpy as np
n=64
A=np.random.rand(n,n )
B=np.random.rand(n,n )
%timeit np.tensordot(A, B)
```
Results
```
main: 23.1 µs ± 488 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
PR: 14.3 µs ± 674 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

See #25713

<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->